[REFACTOR] implemented 2 new services for future authentication method implementations
