+++ molgenis-security/src/main/java/org/molgenis/security/google/GoogleAuthenticatorService.java
@@ -1,0 +1,14 @@
+package org.molgenis.security.google;
+
+/**
+ * @author tommy
+ */
+public interface GoogleAuthenticatorService
+{
+	/**
+	 * <p>Genreate URI for Google Authenticator</p>
+	 *
+	 * @return Google Authenticator URI
+	 */
+	String getGoogleAuthenticatorURI(String secretKey);
+}
