+++ molgenis-security/src/main/java/org/molgenis/security/twofactor/TwoFactorAuthenticationServiceImpl.java
@@ -1,7 +1,6 @@
 package org.molgenis.security.twofactor;
 
 import org.apache.commons.codec.binary.Base32;
-import org.jboss.aerogear.security.otp.Totp;
 import org.molgenis.auth.User;
 import org.molgenis.auth.UserMetaData;
 import org.molgenis.data.DataService;
@@ -19,8 +18,6 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
-import java.io.UnsupportedEncodingException;
-import java.net.URLEncoder;
 import java.security.SecureRandom;
 import java.util.ArrayList;
 import java.util.List;
@@ -33,63 +30,35 @@ import static org.molgenis.security.core.runas.RunAsSystemProxy.runAsSystem;
 public class TwoFactorAuthenticationServiceImpl implements TwoFactorAuthenticationService
 {
 	private AppSettings appSettings;
+	private OTPService otpService;
 	private DataService dataService;
 
-	public TwoFactorAuthenticationServiceImpl(AppSettings appSettings, DataService dataService)
+	public TwoFactorAuthenticationServiceImpl(AppSettings appSettings, OTPService otpService, DataService dataService)
 	{
 		this.appSettings = requireNonNull(appSettings);
+		this.otpService = otpService;
 		this.dataService = requireNonNull(dataService);
 	}
 
 	@Override
-	public boolean isVerificationCodeValid(String verificationCode)
+	public boolean isVerificationCodeValidForUser(String verificationCode)
 			throws BadCredentialsException, UsernameNotFoundException
 	{
-		boolean isValid = true;
+		boolean isValid = false;
 		UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 
 		if (appSettings.getTwoFactorAuthentication().equals(TwoFactorAuthenticationSetting.ENABLED.toString()))
 		{
 			User user = runAsSystem(() -> dataService.findOne(USER,
 					new QueryImpl<User>().eq(UserMetaData.USERNAME, userDetails.getUsername()), User.class));
-			if (StringUtils.hasText(user.getSecretKey()))
+
+			if (otpService.tryVerificationCode(verificationCode, user.getSecretKey()))
 			{
-				if (verificationCode == null)
-				{
-					throw new BadCredentialsException("2 factor authentication code is mandatory");
-				}
-				final Totp totp = new Totp(user.getSecretKey());
-				if (!isValidLong(verificationCode) || !totp.verify(verificationCode))
-				{
-					throw new BadCredentialsException("Invalid 2 factor authentication code");
-				}
+				isValid = true;
 			}
-			else
-			{
-				throw new BadCredentialsException("2 factor authentication secret key is not available");
-			}
+
 		}
 		return isValid;
-	}
-
-	@Override
-	public boolean tryVerificationCode(String verificationCode, String secretKey)
-	{
-		final Totp totp = new Totp(secretKey);
-		return isValidLong(verificationCode) && totp.verify(verificationCode);
-	}
-
-	private boolean isValidLong(String code)
-	{
-		try
-		{
-			Long.parseLong(code);
-		}
-		catch (NumberFormatException e)
-		{
-			return false;
-		}
-		return true;
 	}
 
 	@Override
@@ -107,6 +76,16 @@ public class TwoFactorAuthenticationServiceImpl implements TwoFactorAuthenticati
 		{
 			throw new UsernameNotFoundException("Can't find user: [" + userDetails.getUsername() + "]");
 		}
+	}
+
+	@Override
+	public String generateSecretKey()
+	{
+		SecureRandom random = new SecureRandom();
+		byte[] bytes = new byte[20];
+		random.nextBytes(bytes);
+		Base32 base32 = new Base32();
+		return base32.encodeToString(bytes);
 	}
 
 	@Override
@@ -151,13 +130,6 @@ public class TwoFactorAuthenticationServiceImpl implements TwoFactorAuthenticati
 	}
 
 	@Override
-	public String getUnAuthenticatedUser()
-	{
-		UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-		return userDetails.getUsername();
-	}
-
-	@Override
 	public void authenticate() throws BadCredentialsException
 	{
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
@@ -168,33 +140,6 @@ public class TwoFactorAuthenticationServiceImpl implements TwoFactorAuthenticati
 				updatedAuthorities);
 
 		SecurityContextHolder.getContext().setAuthentication(newAuth);
-	}
-
-	@Override
-	public String generateSecretKey()
-	{
-		SecureRandom random = new SecureRandom();
-		byte[] bytes = new byte[20];
-		random.nextBytes(bytes);
-		Base32 base32 = new Base32();
-		return base32.encodeToString(bytes);
-	}
-
-	@Override
-	public String getGoogleAuthenticatorURI(String secretKey)
-	{
-		String normalizedBase32Key = secretKey.replace(" ", "").toUpperCase();
-		String user = "";
-		try
-		{
-			return "otpauth://totp/" + URLEncoder.encode("molgenis" + ":" + "admin", "UTF-8").replace("+", "%20")
-					+ "?secret=" + URLEncoder.encode(normalizedBase32Key, "UTF-8").replace("+", "%20") + "&issuer="
-					+ URLEncoder.encode("molgenis", "UTF-8").replace("+", "%20");
-		}
-		catch (UnsupportedEncodingException e)
-		{
-			throw new IllegalStateException(e);
-		}
 	}
 
 }
